# Conflicts:
#	src/main/java/com/squad03/flap/controller/EmpresaController.java
#	src/main/java/com/squad03/flap/model/Tarefa.java
#	src/main/resources/application.properties

+++ src/main/java/com/squad03/flap/service/ListaService.java
@@ -1,49 +1,65 @@
 package com.squad03.flap.service;
 
 import com.squad03.flap.DTO.AtualizacaoLista;
+import com.squad03.flap.DTO.BuscaEmpresa;
 import com.squad03.flap.DTO.CadastroLista;
 import com.squad03.flap.DTO.BuscaLista;
 import com.squad03.flap.model.Lista;
 import com.squad03.flap.repository.ListaRepository;
+import jakarta.transaction.Transactional;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Optional;
 import java.util.stream.Collectors;
 
 @Service
 public class ListaService {
 
+    @Autowired
     private ListaRepository repository;
 
-    @Autowired
-    public ListaService(ListaRepository repository) {
-        this.repository = repository;
+    @Transactional
+    public BuscaLista cadastrarLista(CadastroLista cadastroLista) {
+        Lista listaNova = new Lista(cadastroLista);
+        Lista listaSalva = repository.save(listaNova);
+        return new BuscaLista(listaSalva);
     }
 
-    public void cadastrarLista(CadastroLista cadastroLista) {
-        repository.save(new Lista(cadastroLista));
-    }
-
-    public Lista buscarListaPorId(int id) {
-        return repository.findById(id).get();
+    public Optional<BuscaLista> buscarListaPorId(Long id) {
+        return repository.findById(id).map(BuscaLista::new);
     }
 
     public List<BuscaLista> buscarListas() {
         return repository.findAll().stream().map(BuscaLista::new).collect(Collectors.toList());
     }
 
-    public void AtualizarLista(int id, AtualizacaoLista atualizacaoLista) {
-        Lista Lista = repository.findById(id).get();
+    @Transactional
+    public Optional<BuscaLista> atualizarLista(Long id, AtualizacaoLista atualizacaoLista) {
+        return repository.findById(id)
+                .map(lista -> {
 
-        Lista.setNome(atualizacaoLista.nome());
-        Lista.setPosicao(atualizacaoLista.posicao());
-        Lista.setCor(atualizacaoLista.cor());
+                    if (atualizacaoLista.nome() != null) {
+                        lista.setNome(atualizacaoLista.nome());
+                    }
+                    if (atualizacaoLista.posicao() != 0) {
+                        lista.setPosicao(atualizacaoLista.posicao());
+                    }
+                    if (atualizacaoLista.cor() != null) {
+                        lista.setCor(atualizacaoLista.cor());
+                    }
 
-        repository.save(Lista);
+                    return new BuscaLista(repository.save(lista));
+                });
     }
 
-    public void excluirLista(int id) {
-        repository.deleteById(id);
+    @Transactional
+    public boolean excluirLista(Long id) {
+        if (repository.existsById(id)) {
+            repository.deleteById(id);
+            return true;
+        }
+        return false;
     }
 }
